master rbac

+++ app/Http/Controllers/DistrictController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers;
 
 
 use App\Helper\CustomController;
+use App\Helper\Formula;
 use App\Models\City;
 use App\Models\District;
+use Illuminate\Support\Facades\Validator;
 
 class DistrictController extends CustomController
 {
@@ -17,17 +19,39 @@ class DistrictController extends CustomController
 
     public function index()
     {
+        $access = $this->getRoleAccess(Formula::APPMasterDistrict);
         if ($this->request->ajax()) {
-            $data = District::with(['city'])->orderBy('created_at', 'ASC')->get();
+            $data = District::with(['city'])->orderBy('created_at', 'DESC')->get();
             return $this->basicDataTables($data);
         }
-        return view('admin.master-data.district.index');
+        return view('admin.master-data.district.index')
+            ->with([
+                'access' => $access
+            ]);
     }
+
+    private $rule = [
+        'city' => 'required',
+        'name' => 'required',
+    ];
+
+    private $message = [
+        'city.required' => 'kolom kota/kabupaten wajib di isi',
+        'name.required' => 'kolom nama wajib di isi',
+    ];
 
     public function store()
     {
+        $access = $this->getRoleAccess(Formula::APPMasterDistrict);
+        if (!$access['is_granted_create']) {
+            abort(403);
+        }
         if ($this->request->method() === 'POST') {
             try {
+                $validator = Validator::make($this->request->all(), $this->rule, $this->message);
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->with('validator', 'Harap Mengisi Kolom Dengan Benar');
+                }
                 $data_request = [
                     'city_id' => $this->postField('city'),
                     'name' => $this->postField('name'),
@@ -44,9 +68,17 @@ class DistrictController extends CustomController
 
     public function patch($id)
     {
+        $access = $this->getRoleAccess(Formula::APPMasterDistrict);
+        if (!$access['is_granted_update']) {
+            abort(403);
+        }
         $data = District::with(['city'])->findOrFail($id);
         if ($this->request->method() === 'POST') {
             try {
+                $validator = Validator::make($this->request->all(), $this->rule, $this->message);
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->with('validator', 'Harap Mengisi Kolom Dengan Benar');
+                }
                 $data_request = [
                     'city_id' => $this->postField('city'),
                     'name' => $this->postField('name'),
@@ -66,6 +98,10 @@ class DistrictController extends CustomController
 
     public function destroy($id)
     {
+        $access = $this->getRoleAccess(Formula::APPMasterDistrict);
+        if (!$access['is_granted_delete']) {
+            return $this->jsonErrorResponse('cannot access delete perform...');
+        }
         try {
             District::destroy($id);
             return $this->jsonSuccessResponse('success');
