+++ app/Http/Controllers/LocomotiveTypeController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 
 
 use App\Helper\CustomController;
+use App\Helper\Formula;
 use App\Models\LocomotiveType;
+use Illuminate\Support\Facades\Validator;
 
 class LocomotiveTypeController extends CustomController
 {
@@ -16,17 +18,38 @@ class LocomotiveTypeController extends CustomController
 
     public function index()
     {
+        $access = $this->getRoleAccess(Formula::APPMasterLocomotiveType);
         if ($this->request->ajax()) {
-            $data = LocomotiveType::with([])->orderBy('code', 'ASC')->get();
+            $data = LocomotiveType::with([])->orderBy('created_at', 'DESC')->get();
             return $this->basicDataTables($data);
         }
-        return view('admin.master-data.locomotive-type.index');
+        return view('admin.master-data.locomotive-type.index')->with([
+            'access' => $access
+        ]);
     }
+
+    private $rule = [
+        'code' => 'required',
+        'name' => 'required',
+    ];
+
+    private $message = [
+        'code.required' => 'kolom kode wajib di isi',
+        'name.required' => 'kolom nama wajib di isi',
+    ];
 
     public function store()
     {
+        $access = $this->getRoleAccess(Formula::APPMasterLocomotiveType);
+        if (!$access['is_granted_create']) {
+            abort(403);
+        }
         if ($this->request->method() === 'POST') {
             try {
+                $validator = Validator::make($this->request->all(), $this->rule, $this->message);
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->with('validator', 'Harap Mengisi Kolom Dengan Benar');
+                }
                 $data_request = [
                     'code' => $this->postField('code'),
                     'name' => $this->postField('name'),
@@ -43,9 +66,17 @@ class LocomotiveTypeController extends CustomController
 
     public function patch($id)
     {
+        $access = $this->getRoleAccess(Formula::APPMasterLocomotiveType);
+        if (!$access['is_granted_update']) {
+            abort(403);
+        }
         $data = LocomotiveType::findOrFail($id);
         if ($this->request->method() === 'POST') {
             try {
+                $validator = Validator::make($this->request->all(), $this->rule, $this->message);
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->with('validator', 'Harap Mengisi Kolom Dengan Benar');
+                }
                 $data_request = [
                     'code' => $this->postField('code'),
                     'name' => $this->postField('name'),
@@ -62,6 +93,10 @@ class LocomotiveTypeController extends CustomController
 
     public function destroy($id)
     {
+        $access = $this->getRoleAccess(Formula::APPMasterLocomotiveType);
+        if (!$access['is_granted_delete']) {
+            return $this->jsonErrorResponse('cannot access delete perform...');
+        }
         try {
             LocomotiveType::destroy($id);
             return $this->jsonSuccessResponse('success');
