+++ app/Http/Controllers/TrainTypeController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 
 
 use App\Helper\CustomController;
+use App\Helper\Formula;
 use App\Models\TrainType;
+use Illuminate\Support\Facades\Validator;
 
 class TrainTypeController extends CustomController
 {
@@ -16,17 +18,38 @@ class TrainTypeController extends CustomController
 
     public function index()
     {
+        $access = $this->getRoleAccess(Formula::APPMasterTrainType);
         if ($this->request->ajax()) {
             $data = TrainType::with([])->orderBy('code', 'ASC')->get();
             return $this->basicDataTables($data);
         }
-        return view('admin.master-data.train-type.index');
+        return view('admin.master-data.train-type.index')->with([
+            'access' => $access
+        ]);
     }
+
+    private $rule = [
+        'code' => 'required',
+        'name' => 'required',
+    ];
+
+    private $message = [
+        'code.required' => 'kolom kode wajib di isi',
+        'name.required' => 'kolom nama wajib di isi',
+    ];
 
     public function store()
     {
+        $access = $this->getRoleAccess(Formula::APPMasterTrainType);
+        if (!$access['is_granted_create']) {
+            abort(403);
+        }
         if ($this->request->method() === 'POST') {
             try {
+                $validator = Validator::make($this->request->all(), $this->rule, $this->message);
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->with('validator', 'Harap Mengisi Kolom Dengan Benar');
+                }
                 $data_request = [
                     'code' => $this->postField('code'),
                     'name' => $this->postField('name'),
@@ -42,9 +65,17 @@ class TrainTypeController extends CustomController
 
     public function patch($id)
     {
+        $access = $this->getRoleAccess(Formula::APPMasterTrainType);
+        if (!$access['is_granted_update']) {
+            abort(403);
+        }
         $data = TrainType::findOrFail($id);
         if ($this->request->method() === 'POST') {
             try {
+                $validator = Validator::make($this->request->all(), $this->rule, $this->message);
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->with('validator', 'Harap Mengisi Kolom Dengan Benar');
+                }
                 $data_request = [
                     'code' => $this->postField('code'),
                     'name' => $this->postField('name'),
@@ -60,6 +91,10 @@ class TrainTypeController extends CustomController
 
     public function destroy($id)
     {
+        $access = $this->getRoleAccess(Formula::APPMasterTrainType);
+        if (!$access['is_granted_delete']) {
+            return $this->jsonErrorResponse('cannot access delete perform...');
+        }
         try {
             TrainType::destroy($id);
             return $this->jsonSuccessResponse('success');
